requirements.txt
Flask==2.2.5
Flask-MongoEngine==1.0.0
celery[redis]==5.3.0
redis==5.0.0